Enhance Cache interface, better TLB fix and begin DRAM fix.
The accessSingleLine function will also provide the writeback adresses
to keep the interface constistent with insertion, and will assert if
the required writeback pointers aren't provided.

TLB for the sake of robustness handle those eviction as it would for
an insertion, but this code should not be run.

DRAM cache will also be treated as not compressible.